Moved tests outside package
Tests are now in a directory called tests in the root directory.
Trying to keep application code seperate from tests, so I also made
a setup.py file, which makes the application accessible to the tests.
Probably will not work, but I'm just giving it a try at this point.
Also, changing my tests by eliminating the conftest.py file because
almost all of the fixtures in that file were local to specific modules.
Now, trying to keep test instances modularized by including module-specific
fixtures in their respective modules and limiting scope to module to
reduce namespace clutter. Also going to try 'parameterizing' test fuctions
with @pytest.mark.parameterize(). This could reduce dependancy on fixtures.
If that does not work or is too messy, I will just include the test
instances/data in the test function definitions. This might be the best
option, but I want to familiarize myself with pytest more by using
their features.