Moved test_board.py fixtures into funcs
Not worth it to have all these random fixtures lying around.
Test files are already cluttered and disorganized. Having test
cases local to the function seems like the best option for right
now. I think fixtures are meant for retrieving large amounts of
data from a database to test against and not for storing a few
tuples of test cases. Maybe later I will try making seperate folders
for each test file, where all of the test cases are included
in other files local to that directory.
pytest-datadir or pytest-datafiles from PyPi might be a
good option, although I have not read the documentation, so
I don't know if these are meant for large data files or test cases
(or both?).