Edited test for Player.longest_road()
Test for Player.longest_road() was not passing. Many issues, including
bad syntax, mutating and iterable during its iteration and poor
algorithm design. Algorithm still does not pass all the tests
but is in better condition than previously, which is why I am
comitting now. I intend to change it so that it counts the number
of consecutive intersections (and subtracts one) instead of
edges because finding the longest path using edges has a lot of
problems that I did not forsee. Intersections should be more
straightforward.