Changed Player.longest_road()
I have changed the longest road algorithm to use intersections
instead of edges to find the longest road. It works better than
before, but I still need to work out some kinks. For example, simply
collecting the coordinates of the intersections which have roads
connected to them is not an effective strategy because two
intersections which are not actually connected by a road but
happen to have roads connected to them elsewhere are added to
the set of 'road intersections'. Because they are adjacent in this
set, the algorithm thinks that they are connected by a road because
they are adjacent, when in fact they are not. Another issue with
this algorithm is the way that it counts length. Roads which loop
back on themselves one or more times have more intersections to
roads. I will fix these problems soon, but wanted to commit changes
because using intersections to count roads is in  fact a step in
the right direction.