Change Player.longest_road()
Trying out different methods for finding the longest road. There
are so many! Was originally trying to append the road paths to a
list when the function satisfies a terminal condition, but things
got complicated when I defined the helper recursion function
_exhaust_pathways outside of the main function. Do not want to
append to a list that is not locally defined in the function
itself, so trying to figure out ways around. Thinking about simply
returing the value instead of appending, but as the roads
bifurcate, one call to _exhaust_pathways could invoke multiple
returns. Not sure what kind of behaviour that would cause, thus the
branch name "player_experiment". Will try other methods, too.