Add end cond _exhaust_pathways
Added an end condition to _exhaust_pathways. If the current intersection has a
structure (settlement or city) on it that is not owned by the player, then it
will terminate. Roads should not pass through the structures of other players.
If they do, then it does not count as a consecutive road.